improvement: backbeat as a lib
Moved config code into lib.
Moved index.js into bin/backbeat.js
Created an index.js that references lib code.
