+++ extensions/lifecycle/management.js
@@ -1,7 +1,7 @@
 const AWS = require('aws-sdk');
 const werelogs = require('werelogs');
 
-const config = require('../../conf/Config');
+const config = require('../../lib/Config');
 const management = require('../../lib/management/index');
 
 const logger = new werelogs.Logger('mdManagement:lifecycle');
