+++ index.js
@@ -1,50 +1,16 @@
-'use strict'; // eslint-disable-line strict
-
-const werelogs = require('werelogs');
-
-const runServer = require('./lib/api/BackbeatServer');
-const { initManagement } = require('./lib/management/index');
-const setupIngestionSiteMock = require('./tests/utils/mockIngestionSite');
-const config = require('./conf/Config');
-const Logger = werelogs.Logger;
-const log = new Logger('BackbeatServer:index');
-
-werelogs.configure({
-    level: config.log.logLevel,
-    dump: config.log.dumpLevel,
-});
-
-function initAndStart() {
-    const repConfig = config.extensions.replication;
-    const sourceConfig = repConfig.source;
-    initManagement({
-        serviceName: 'replication',
-        serviceAccount: sourceConfig.auth.account,
-        enableIngestionUpdates: true,
-    }, error => {
-        if (error) {
-            log.error('could not load management db', { error });
-            setTimeout(initAndStart, 5000);
-            return;
-        }
-        log.info('management init done');
-
-        const bootstrapList = config.getBootstrapList();
-        repConfig.destination.bootstrapList = bootstrapList;
-
-        config.on('bootstrap-list-update', () => {
-            repConfig.destination.bootstrapList = config.getBootstrapList();
-        });
-
-        runServer(config, Logger);
-    });
-}
-
-if (process.env.CI === 'true') {
-    // skip initManagement
-    // set mock config ingestion site on start-up
-    setupIngestionSiteMock();
-    runServer(config, Logger);
-} else {
-    initAndStart();
-}
+module.exports = {
+    Config: require('./lib/Config'),
+    management: require('./lib/management/index'),
+    BackbeatConsumer: require('./lib/BackbeatConsumer'),
+    BackbeatProducer: require('./lib/BackbeatProducer'),
+    clients: {
+        zookeeper: require('./lib/clients/zookeeper'),
+    },
+    models: {
+        QueueEntry: require('./lib/models/QueueEntry'),
+        ObjectQueueEntry: require('./lib/models/ObjectQueueEntry'),
+    },
+    credentials: {
+        AccountCredentials: require('./lib/credentials/AccountCredentials'),
+    },
+};
