+++ lib/Config.js
@@ -33,7 +33,7 @@ class Config extends EventEmitter {
          * By default, the config file is "config.json" at the root.
          * It can be overridden using the BACKBEAT_CONFIG_FILE environment var.
          */
-        this._basePath = __dirname;
+        this._basePath = `${__dirname}/../conf`;
         if (process.env.BACKBEAT_CONFIG_FILE !== undefined) {
             this._configPath = process.env.BACKBEAT_CONFIG_FILE;
         } else {
