+++ lib/clients/VaultClientCache.js
@@ -1,4 +1,4 @@
-const config = require('../../conf/Config');
+const config = require('../Config');
 
 const VaultClient = require('vaultclient').Client;
