+++ lib/config.joi.js
@@ -2,7 +2,7 @@
 
 const joi = require('@hapi/joi');
 const { hostPortJoi, transportJoi, logJoi, certFilePathsJoi } =
-      require('../lib/config/configItems.joi.js');
+      require('./config/configItems.joi.js');
 
 const logSourcesJoi = joi.string().valid('bucketd', 'mongo', 'ingestion',
     'dmd');
