+++ lib/management/index.js
@@ -1,7 +1,7 @@
 const async = require('async');
 const forge = require('node-forge');
 const arsenal = require('arsenal');
-const config = require('../../conf/Config');
+const config = require('../Config');
 const werelogs = require('werelogs');
 const bucketclient = require('bucketclient');
