+++ lib/management/patchConfiguration.js
@@ -2,7 +2,7 @@
 
 const async = require('async');
 
-const config = require('../../conf/Config');
+const config = require('../Config');
 
 function updateIngestionBuckets(locations, metadata, logger, cb) {
     metadata.getIngestionBuckets(logger, (err, buckets) => {
