+++ lib/queuePopulator/IngestionPopulator.js
@@ -5,7 +5,7 @@ const schedule = require('node-schedule');
 
 const Logger = require('werelogs').Logger;
 
-const config = require('../../conf/Config');
+const config = require('../Config');
 const IngestionReader = require('./IngestionReader');
 const BackbeatProducer = require('../BackbeatProducer');
 const MetricsConsumer = require('../MetricsConsumer');
