+++ lib/queuePopulator/LogReader.js
@@ -2,7 +2,7 @@ const async = require('async');
 
 const jsutil = require('arsenal').jsutil;
 
-const config = require('../../conf/Config');
+const config = require('../Config');
 const BackbeatProducer = require('../BackbeatProducer');
 const ReplicationQueuePopulator =
     require('../../extensions/replication/ReplicationQueuePopulator');
