+++ package.json
@@ -24,7 +24,7 @@
     "bh_test": "mocha --recursive tests/behavior",
     "lint": "eslint $(git ls-files '*.js')",
     "lint_md": "mdlint $(git ls-files '*.md')",
-    "start": "node index.js"
+    "start": "node bin/backbeat.js"
   },
   "repository": {
     "type": "git",
