+++ tests/unit/conf/Config.js
@@ -1,7 +1,7 @@
 'use strict'; // eslint-disable-line
 
 const assert = require('assert');
-const config = require('../../../conf/Config');
+const config = require('../../../lib/Config');
 
 describe('backbeat config parsing and validation', () => {
     it('should parse correctly the default config', () => {
